requirements.txt
fastapi
uvicorn
numpy
python-dotenv